Fix merge regressions in propagation and spectroscopy support
Resolve unbounded Krylov growth, happy-breakdown handling, block-Arnoldi residuals, thread-local adaptive state, pulse-sequence indexing, case-sensitive identifiers, empty semiclassical matrices, and six-component strain mappings.

Restore reliable CMake/CTest integration, remove generated test build files and compiler warnings, and add regression coverage for propagation, pulses, strain, and semiclassical Hamiltonians.

Integrate the remote selective-test controls without allowing zero-test or stale-object runs, restore optimized Makefile builds, keep fixed-dimension Krylov calls non-adaptive by default, and compare instant-pulse SS/HS trajectories at matching physical times.

Preserve the established RWA and secular spectroscopy behavior. SS/HS EPR, CIDNP, MARY, all five bRC_CISS spectra, and a full-grid 432-state dmCry point agree with time_resolved_vs_cw_spectra; CMake/CTest and the Makefile pass all 127 tests, and the focused powder target passes all 16 selected tests.

+++ RunSection/Tasks/TaskMultiStaticSSTimeEvo.cpp
@@ -382,7 +382,7 @@ namespace RunSection
 		{
 			// Propagate
 			//evaluate pulse sequence 
-			int previous_current_block = current_block;
+			size_t previous_current_block = current_block;
 			ClampTimeEvolution(CurrentTime, this->totaltime, time_blocks, current_block, this->timestep, params);
 			if(current_block != previous_current_block)
 				this->timestep = timestep / 100;
